+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.0.5+5
+* invert offset value
+* add proper example
+
+# 0.0.4+4
+# 0.0.3+3
+* Refactor
+
 # 0.0.2+2
 * Replace 'header' with 'headers'